:recycle: Split radio-select component into its own module
